+++ src/pke/include/scheme/cryptocontextparams-defaults.h
@@ -76,8 +76,8 @@ constexpr ExecutionMode executionMode                     = EXEC_EVALUATION;
 constexpr DecryptionNoiseMode decryptionNoiseMode         = FIXED_NOISE_DECRYPT;
 constexpr double noiseEstimate                            = 0;
 constexpr double desiredPrecision                         = 25;
-constexpr double statisticalSecurity                      = 30;
-constexpr double numAdversarialQueries                    = 1;
+constexpr uint32_t statisticalSecurity                    = 30;
+constexpr uint32_t numAdversarialQueries                  = 1;
 };  // namespace CKKSRNS_SCHEME_DEFAULTS
 
 namespace BFVRNS_SCHEME_DEFAULTS {
@@ -115,8 +115,8 @@ constexpr ExecutionMode executionMode                     = EXEC_EVALUATION;
 constexpr DecryptionNoiseMode decryptionNoiseMode         = FIXED_NOISE_DECRYPT;
 constexpr double noiseEstimate                            = 0;
 constexpr double desiredPrecision                         = 0;
-constexpr double statisticalSecurity                      = 30;
-constexpr double numAdversarialQueries                    = 1;
+constexpr uint32_t statisticalSecurity                    = 30;
+constexpr uint32_t numAdversarialQueries                  = 1;
 };  // namespace BFVRNS_SCHEME_DEFAULTS
 
 namespace BGVRNS_SCHEME_DEFAULTS {
@@ -146,8 +146,8 @@ constexpr ExecutionMode executionMode                     = EXEC_EVALUATION;
 constexpr DecryptionNoiseMode decryptionNoiseMode         = FIXED_NOISE_DECRYPT;
 constexpr double noiseEstimate                            = 0;
 constexpr double desiredPrecision                         = 0;
-constexpr double statisticalSecurity                      = 30;
-constexpr double numAdversarialQueries                    = 1;
+constexpr uint32_t statisticalSecurity                    = 30;
+constexpr uint32_t numAdversarialQueries                  = 1;
 };  // namespace BGVRNS_SCHEME_DEFAULTS
 
 //====================================================================================================================
